requirements.txt
torch>=2.0.0
torchvision>=0.15.0
scikit-learn>=1.3.0
modAL-python>=0.4.1
streamlit>=1.25.0
pillow>=10.0.0
pandas>=2.0.0
joblib>=1.3.0
numpy>=1.24.0